feat(streaming): add notify
Add notification functionality to allow a signal to be sent when
the call is made to receive audio data. This allows the user to
react to this event and send data when appropiate. This hopefully
prevents audio data being sent from the user before the stream is ready,
resulting in the streaming breaking.