chore(karma): set transport to polling
It seems that autobahn conflicts with karma in the browser. Autobahn
provides websocket functionality (though its own means) and karma uses
the default socket.io protocol for this.

By setting the transport protocol to "polling" (yes, back to the
future!) the problems with failing build should be history for now.