fix(VolumeMeter): be able to break the analyzing loop
The animation loop somehow could not break out of its loop when
needed. By taking another approach to access the value for allowing
the loop ('willAnimate') things got better.

Besides setting the analyzer to stop also a resume function has been
added to be able to reuse the VolumeMeter component.