refactor: begin the transition from classes to functions.
Refactor and extend the communication mechanism with the API to be more
dynamic/flexible. This makes it also to doe (authorised) PUT requets.
Also convert the authentication mechanism to functions using the new
communication mechanism.

Update the README.md to contain required information. It is still
incomplete; the rest will follow during the refactor.