chore: properly babelify the source
UMD seems to be the proper way to expose data; all other forms provide a
couple of quirks; such as common.js (the default of babel) which only
allows you to import exposed objects when you specify the file which
exposes it, including the index.js of a module (e.g.
`import {a} from 'foo/bar/index.js';`). UMD doesn't require this.

Also rename the `babelify` npm script to `build`; because that is a more
sane thing to do.