improve packaging
- version bump of all used packages
- removal of concurrency, no longer used
- removal of bufferutil, not used at all
- set engines versions to the correct versions
- add packages folder

The packages folder will container a sort of boilerplate for the build
(that will be published). In there you'll find a package.json, LICENSE
and README.md.

The package.json is almost the same as the one in the root. Exept for no
scripts and no devDependencies (the're not needed for now).

Note that version bumps for deps in the root package.json will also need
to take place in the packages folder.