do not use es6 promises for autobahn
Autobahn gives an option to use es6 Promises in stead of their default
When.JS implementation. Well, that'll break stuff. If we would use
progressed results (and we do that) there needs to be support for notify
and progress. When.JS did support that (but it is deprecated for now).

So, TL;DR: disable es6 promises in favor of the default autobahn choice.