improve github pages deploy process
Make use of a script to deploy github pages. The change is important:
the pages provider from travis-ci (which works awesome) relies on using
a personal access token. This token could be used to also get access to
other repositories of the token owner. We don't want that!

So, also added a ney deploy key, added a script, added the appropiate
settings in the .travis.yml and last but not least: added deploy-key
private and public to the .gitignore file!

The script has now been made to do the following: if it is master, than
update the root of the gh-pages branch. If it is a tag (with the vX.X.X
pattern!) then the docs will be pushed to a specific versions folder.