improve travis build yet again
- Remove the deploy key and corresponding settings in our deploy script.
We try to use personal access tokens for all, but as our itslanguage bot
user.

- Script determines the https url to the repo. This is needed because
this is the only way personal access tokens work.

- Push got some extra parameters to make sure no sensitive data leaks to
places where it shouldn't go.

- All branches that trigger the script get their own folder in the
gh-pages branch. This means that there needs to be a general index.html
for our pages (not here yet).

- Our .travis.yml uses after_success to do the build. This was in the
before_deploy first, but if it fails, the script continues. Now it
shouldn't anymore.

- Our .travis.yml file now installs the latest yarn version.

- In general tried to add some more documentation comments around.