Fix allow itslanguage urls (#303)
fix: make use of the correct apiUrl setting

The SDK made use of apiURL and apiUrl on the settings object. Off
course, these two are not the same at runtime. So changed it to use a
single one: apiUrl.

Also add missing dependency to release package (debug package was
missing).