refactor: user api
Refactored all user functions to be simpeler: create, getCurrent, getAll
and getById instead of creatUser, getAllUsers, etc. This makes it in SDK
terms more easy to use.

Added the getCurrent function to get the current user. For this also
another URL was added (since this feature is on another url endpoint).