chore: improve request method
Removed the truthy check on url and replaced it for an falsy check. If
the url is not set, the default will be used. At run time this will lead
to errors (which is okay).

Also set the correct Content-Type header for different types of requests
. Using URLSearchParams as body should lead to
'application/x-www-form-urlencoded'. FormData should lead to
'multipart/form-data' and an object to 'application/json'. All else to
'text/plain'.

In case of URLSearchParams, we use toString() to get the correct format
to send.

Oh, the test also adheres to the changes and some comments where added
for clarification.