fix: posting FormData files
There was an error in the request for posting to an URL with FormData.
By "forcing" the `Content-Type: multipart/form-data` header we omitted
the needed boundary information. The browser and/or some NodeJS process
will set this automatically when you try to post FormData: so there is
no need to specify it here this way.

- Do not set default content-type to be able to not set the content-type
  when FormData is being posted.

- Do not set multipart/form-data when sending FormData. It is missing
  the essential boundary information now, which will automatically be
  set by either the browser or some NodeJS application.