refactor: audio-over-socket.spec.js tests
It seems the async/await way of doing the test causes them to sometimes
fail because some other test messes up. Messes up is due to using some
variables that get overridden multiple times making them unreliable.

Moved this to a more solid approach and put it all in a setup function.